Initial colors
