+++ packages/app/features/home/screen.tsx
@@ -1,4 +1,4 @@
-import { Anchor, Button, H1, Paragraph, Separator, Sheet, XStack, YStack } from '@my/ui'
+import { Button, H1, Paragraph, Separator, Sheet, Text, XStack, YStack } from '@my/ui'
 // import { ChevronDown, ChevronUp } from '@tamagui/lucide-icons'
 // import { Entypo } from '@expo/vector-icons'
 import React, { useState } from 'react'
@@ -10,80 +10,22 @@ export function HomeScreen() {
   })
 
   return (
-    <YStack f={1} jc="center" ai="center" p="$4" space>
+    <YStack f={1} jc="center" ai="center" p="$4" space backgroundColor="$haiti">
       <YStack space="$4" maw={600}>
-        <H1 ta="center">Welcome to Tamagui.</H1>
-        <Paragraph ta="center">
-          Here's a basic starter to show navigating from one screen to another. This screen uses the
-          same code on Next.js and React Native.
-        </Paragraph>
-
-        <Separator />
-        <Paragraph ta="center">
-          Made by{' '}
-          <Anchor color="$color12" href="https://twitter.com/natebirdman" target="_blank">
-            @natebirdman
-          </Anchor>
-          ,{' '}
-          <Anchor
-            color="$color12"
-            href="https://github.com/tamagui/tamagui"
-            target="_blank"
-            rel="noreferrer"
-          >
-            give it a ⭐️
-          </Anchor>
+        <H1 ta="center" color="$moonRaker">
+          Arc
+        </H1>
+        <Separator borderColor="$blueBellFaded" />
+        <Paragraph ta="center" color="$blueBell">
+          The Worst Chat App
         </Paragraph>
       </YStack>
 
-      <XStack>
-        <Button {...linkProps}>Link to user</Button>
+      <XStack mt="$8">
+        <Button size="$6" backgroundColor="$electricIndigo" {...linkProps}>
+          Get started
+        </Button>
       </XStack>
-
-      <SheetDemo />
     </YStack>
-  )
-}
-
-function SheetDemo() {
-  const [open, setOpen] = useState(false)
-  const [position, setPosition] = useState(0)
-  return (
-    <>
-      <Button
-        size="$6"
-        // icon={
-        //   open ? (
-        //     <Entypo name="chevron-down" size={24} color="black" />
-        //   ) : (
-        //     <Entypo name="chevron-up" size={24} color="black" />
-        //   )
-        // }
-        circular
-        onPress={() => setOpen((x) => !x)}
-      />
-      <Sheet
-        modal
-        open={open}
-        onOpenChange={setOpen}
-        snapPoints={[80]}
-        position={position}
-        onPositionChange={setPosition}
-        dismissOnSnapToBottom
-      >
-        <Sheet.Overlay />
-        <Sheet.Frame ai="center" jc="center">
-          <Sheet.Handle />
-          <Button
-            size="$6"
-            circular
-            // icon={<Entypo name="chevron-down" size={24} color="black" />}
-            onPress={() => {
-              setOpen(false)
-            }}
-          />
-        </Sheet.Frame>
-      </Sheet>
-    </>
   )
 }
